+++ content.js
@@ -43,11 +43,13 @@ function addListeners(events) {
 
   toggleListener(window, 'copy', onEvent, types.includes('window.copy'));
   toggleListener(window, 'cut',  onEvent, types.includes('window.cut'));
+  toggleListener(window, 'compositionstart', onEvent, types.includes('window.compositionstart'));
 }
 
 function removeListeners() {
   window.removeEventListener('copy', onEvent);
   window.removeEventListener('cut', onEvent);
+  window.removeEventListener('compositionstart', onEvent);
 }
 
 globalThis.requestIdleCallback(() => {