+++ CHANGES.md
@@ -1,6 +1,10 @@
 CHANGES
 =======
 
+## 3.x.x (?)
+
+* Add compositionstart event.
+
 ## 3.2.0 (2025-06-20)
 
 * Support various Tab events: opened / closed / attached / pin / unpin.